Count molecules per region of interest (e.g., cell)

This function takes the information from the molecules and boundaries slot, and counts the molecules per region of interest. Its input is a MoleculeExperiment object, and its output a SpatialExperiment object. That way, if one is interested in doing downstream analyses at the cell level, one can do so.

countMolecules(
  me,
  boundaries_assay = "cell",
  segmentation_info = "boundaries",
  molecules_assay = "detected"
)

Arguments

me

MoleculeExperiment object containing both the transcript data as well as the boundaries data. I.e., the "molecules" and "boundaries" slots need to be filled. See MoleculeExperiment() for more information.

boundaries_assay

Character string naming the list of the boundaries slot from which boundary information should be retrieved from. For example, for counting transcripts per cell, the list containing the cell boundaries (e.g., "cell") should be selected.

segmentation_info

Character string specifying the type of segmentation information available. Can be either "boundaries" or "masks". Currently, only the "boundaries" information is supported.

molecules_assay

Character string naming the list of the molecules slot from which transcript information should be retrieved from. The default is the detected transcript data that is read in when creating a MoleculeExperiment object. It is possible to change it to another mode, e.g., "high_threshold" will access the transcript information that has been stored in the "high_threshold" element of the list in the molecules slot.

Value

A SpatialExperiment object derived from a MoleculeExperiment object.

Examples

repo_dir <- system.file("extdata", package = "MoleculeExperiment")

me <- readXenium(repo_dir,
                  n_samples = 2,
                  keep_cols = "essential")
#> 
#> Detected transcript information can be accessed with molecules(me) or
#> molecules(me, "detected")
#> Boundary information can be accessed with boundaries(me)

spe <- countMolecules(me)
#> The boundaries were retrieved from the "cell" assay. A
#> different assay (e.g., "nucleus") can be specified in the boundaries_assay
#> argument)
#> The transcripts from the detected assay were
#> retrieved. Other assay transcripts can be retrieved by specifying the assay_name
#> argument.
#> The transcripts from the detected assay were
#> retrieved. Other assay transcripts can be retrieved by specifying the assay_name
#> argument.
spe
#> class: SpatialExperiment 
#> dim: 178 14 
#> metadata(0):
#> assays(1): counts
#> rownames(178): 2010300C02Rik Acsbg1 ... Trem2 Trp73
#> rowData names(0):
#> colnames(14): Xenium_V1_FF_Mouse_Brain_MultiSection_1_outs.67500
#>   Xenium_V1_FF_Mouse_Brain_MultiSection_1_outs.67512 ...
#>   Xenium_V1_FF_Mouse_Brain_MultiSection_2_outs.65070
#>   Xenium_V1_FF_Mouse_Brain_MultiSection_2_outs.65071
#> colData names(4): sample_id x_location y_location cell_id
#> reducedDimNames(1): spatial
#> mainExpName: NULL
#> altExpNames(0):
#> spatialCoords names(2) : x_location y_location
#> imgData names(0):