Read in Cosmx data (Nanostring) as an ME object.

This function is a wrapper around the readMolecules function. Note that it can currently only create a simple ME object with the molecules slot filled. Boundary information is not handled yet.

Usage

readCosmx(dataDir, keepCols = "essential", addBoundaries = "cell")

Arguments

dataDir

Character string specifying the directory with the Cosmx output files.

keepCols

Character string specifying which columns to keep. Defaults to "essential". The other option is to select "all", or custom columns by specifying their names in a vector.

addBoundaries

A string with which to specify the name of the boundary assay to be added to the me object. Can be a string, or NULL. If NULL, a simple ME object with no boundaries will be created.

Value

A MoleculeExperiment object

Examples

repoDir <- system.file("extdata", package = "MoleculeExperiment")
repoDir <- paste0(repoDir, "/nanostring_Lung9_Rep1")
#
meCosmx <- readCosmx(repoDir,
    keepCols = "essential"
)
#> ⠙ 1/3 Reading transcripts
#> ✔ 1/3 Reading transcripts [47ms]
#> 
#> ⠙ 2/3 Transforming masks into polygons
#> ✔ 2/3 Transforming masks into polygons [75ms]
#> 
#> ⠙ 3/3 Merging patches
#> ✔ 3/3 Merging patches [40ms]
#> 
meCosmx
#> class:  MoleculeExperiment 
#> 2 samples: sample_1 sample_2 
#> 
#> @molecules contents: 
#> -detected assay:
#> 977 unique features across all samples in assay detected: AATK ABL1 ABL2 ACE
#> ACE2 ACKR1...
#> 15499.5 molecules on average across all samples in assay detected
#> Location range across all samples in assay "detected": [924.01,3002] x [26290,26398]
#> 
#> @boundaries contents:
#> -cell:
#> 98 unique segment IDs: 1 2 3 4 5 6 ...