Read in Merscope data to an ME object

Reads in Merscope (Vizgen) molecule and boundary data from a directory, and standardises it into a MoleculeExperiment object.

Usage

readMerscope(dataDir, keepCols = "essential", addBoundaries = "cell")

Arguments

dataDir

Character string specifying the directory with the Cosmx output files.

keepCols

Vector of characters specifying the columns of interest from the transcripts file. "essential" selects columns with gene names, x and y locations. "all" will select all columns. Alternatively, specific colums of interest can be selected by specifying them as characters in a vector. Note that this personalised vector needs to contain the essential columns.

addBoundaries

A string with which to specify the name of the boundary assay to be added to the me object. Can be a string, or NULL. If NULL, a simple ME object with no boundaries will be created.

Value

A MoleculeExperiment object

Examples

repoDir <- system.file("extdata", package = "MoleculeExperiment")
repoDir <- paste0(repoDir, "/vizgen_HumanOvarianCancerPatient2Slice2")
meMerscope <- readMerscope(repoDir,
    keepCols = "essential",
    addBoundaries = "cell"
)
#> ⠙ 1/2 Reading transcripts
#> ✔ 1/2 Reading transcripts [46ms]
#> 
meMerscope
#> class:  MoleculeExperiment 
#> 1 samples: vizgen_HumanOvarianCancerPatient2Slice2 
#> 
#> @molecules contents: 
#> -detected assay:
#> 486 unique features across all samples in assay detected: ACKR3 ACTA2 ADAMTS4
#> AKT1 AKT2 AKT3...
#> 15160 molecules on average across all samples in assay detected
#> Location range across all samples in assay "detected": [10219.02,10386.83] x [8350.93,8395.3]
#> 
#> @boundaries contents:
#> -cell:
#> 24 unique segment IDs: 45862 45865 45872 45873 45882 45887 ...