Convert a transcripts or boundaries file to the ME list format

The goal of this function is to standardise transcripts and boundaries files for input to a MoleculeExperiment object.

Usage

dataframeToMEList(
  df,
  df_type = NULL,
  assay_name = NULL,
  sample_col = "sample_id",
  factor_col,
  x_col = "x_location",
  y_col = "y_location",
  keep_cols = "essential",
  scale_factor = 1
)

Arguments

df

A data.frame containing the transcripts information or the boundaries information. NOTE: this dataframe should, at a minimum, have the following 4 columns: sample_id, factor_col (e.g., feature_name in transcripts, or cell_id in boundaries), x_location and y_location.

df_type

Character string specifying contents of the dataframe. Can be either "transcripts" or "boundaries".

assay_name

Name with which to identify the information later on in an ME object.

sample_col

Character string specifying the name of the column with the sample id.

factor_col

Character string specifying the name of the column with the factors with which to group the data in the lists. When working with molecules, this column would be e.g., "feature_name" in xenium. When working with boundaries, this column would be e.g., "cell_id" in xenium.

x_col

Character string specifying the name of the column with global x coordinates.

y_col

Character string specifying the name of the column with global y coordinates.

keep_cols

Character string which can be either "essential" or "all". If "essential", the function will only work with the x and y location information.

scale_factor

Integer specifying the scale factor by which to change the scale of the x and y locations (e.g., to change from pixel to micron). The default value is 1.

Value

A list with the format required to input it into slots of a MoleculeExperiment object.

Examples

molecules_df <- data.frame(
    sample_id = rep(c("sample1", "sample2"), times = c(30, 20)),
    features = rep(c("gene1", "gene2"), times = c(20, 30)),
    x_coords = runif(50),
    y_coords = runif(50)
)

molecules_ls <- dataframeToMEList(molecules_df,
                                  df_type = "transcripts",
                                  assay_name = "detected",
                                  sample_col = "sample_id",
                                  factor_col = "features",
                                  x_col = "x_coords",
                                  y_col = "y_coords")

molecules_ls
#> $detected
#> $detected$sample1
#> $detected$sample1$gene1
#> # A tibble: 20 × 2
#>    x_location y_location
#>         <dbl>      <dbl>
#>  1     0.938    0.527   
#>  2     0.943    0.0920  
#>  3     0.905    0.468   
#>  4     0.270    0.545   
#>  5     0.921    0.103   
#>  6     0.0450   0.685   
#>  7     0.681    0.947   
#>  8     0.748    0.718   
#>  9     0.426    0.0826  
#> 10     0.938    0.381   
#> 11     0.309    0.974   
#> 12     0.482    0.000232
#> 13     0.124    0.980   
#> 14     0.327    0.797   
#> 15     0.674    0.186   
#> 16     0.897    0.148   
#> 17     0.643    0.333   
#> 18     0.0827   0.00714 
#> 19     0.714    0.0984  
#> 20     0.814    0.0193  
#> 
#> $detected$sample1$gene2
#> # A tibble: 10 × 2
#>    x_location y_location
#>         <dbl>      <dbl>
#>  1     0.189       0.986
#>  2     0.0486      0.737
#>  3     0.193       0.608
#>  4     0.896       0.226
#>  5     0.428       0.799
#>  6     0.520       0.645
#>  7     0.841       0.890
#>  8     0.0480      0.131
#>  9     0.525       0.871
#> 10     0.107       0.616
#> 
#> 
#> $detected$sample2
#> $detected$sample2$gene2
#> # A tibble: 20 × 2
#>    x_location y_location
#>         <dbl>      <dbl>
#>  1     0.237       0.130
#>  2     0.0361      0.903
#>  3     0.337       0.178
#>  4     0.203       0.955
#>  5     0.384       0.314
#>  6     0.539       0.486
#>  7     0.854       0.173
#>  8     0.915       0.243
#>  9     0.0876      0.419
#> 10     0.604       0.854
#> 11     0.342       0.430
#> 12     0.932       0.735
#> 13     0.598       0.784
#> 14     0.0300      0.843
#> 15     0.298       0.353
#> 16     0.685       0.912
#> 17     0.265       0.127
#> 18     0.902       0.918
#> 19     0.439       0.172
#> 20     0.695       0.408
#> 
#> 
#>