Convert a transcript (molecule) or boundary dataframe to the ME list format

The goal of this function is to standardise transcript and boundary files for input to a MoleculeExperiment object.

Usage

dataframeToMEList(
  df,
  dfType = NULL,
  assayName = NULL,
  sampleCol = "sample_id",
  factorCol = NULL,
  xCol = "x_location",
  yCol = "y_location",
  keepCols = "essential",
  scaleFactor = 1
)

Arguments

df

A data.frame containing the transcript information or the boundary information. NOTE: this dataframe should, at a minimum, have the following 4 columns: sample_id, factorCol (e.g., feature_id in transcripts, or cell_id in boundaries), x_location and y_location.

dfType

Character string specifying contents of the dataframe. Can be either "molecules" or "boundaries".

assayName

Character string specifying the name with which to identify the information later on in an ME object.

sampleCol

Character string specifying the name of the column with the sample id.

factorCol

Character string specifying the name of the column with the factors with which to group the data in the lists. When working with molecules, this column would be e.g., "feature_id" in xenium. When working with boundaries, this column would be e.g., "cell_id" in xenium.

xCol

Character string specifying the name of the column with global x coordinates.

yCol

Character string specifying the name of the column with global y coordinates.

keepCols

Character string which can be either "essential" or "all". If "essential", the function will only work with the x and y location information.

scaleFactor

Integer specifying the scale factor by which to change the scale of the x and y locations (e.g., to change from pixel to micron). The default value is 1.

Value

A list with the format required to input it into slots of a MoleculeExperiment object.

Examples

moleculesDf <- data.frame(
    sample_id = rep(c("sample1", "sample2"), times = c(30, 20)),
    features = rep(c("gene1", "gene2"), times = c(20, 30)),
    x_coords = runif(50),
    y_coords = runif(50)
)

moleculesMEList <- dataframeToMEList(moleculesDf,
                                  dfType = "molecules",
                                  assayName = "detected",
                                  sampleCol = "sample_id",
                                  factorCol = "features",
                                  xCol = "x_coords",
                                  yCol = "y_coords")

moleculesMEList
#> $detected
#> $detected$sample1
#> $detected$sample1$gene1
#> # A tibble: 20 × 2
#>    x_location y_location
#>         <dbl>      <dbl>
#>  1      0.799     0.776 
#>  2      0.465     0.322 
#>  3      0.787     0.891 
#>  4      0.810     0.114 
#>  5      0.444     0.137 
#>  6      0.234     0.827 
#>  7      0.652     0.0262
#>  8      0.762     0.520 
#>  9      0.628     0.896 
#> 10      0.376     0.884 
#> 11      0.603     0.415 
#> 12      0.655     0.309 
#> 13      0.515     0.256 
#> 14      0.657     0.439 
#> 15      0.545     0.725 
#> 16      0.956     0.612 
#> 17      0.831     0.191 
#> 18      0.111     0.213 
#> 19      0.228     0.154 
#> 20      0.265     0.650 
#> 
#> $detected$sample1$gene2
#> # A tibble: 10 × 2
#>    x_location y_location
#>         <dbl>      <dbl>
#>  1     0.398       0.893
#>  2     0.952       0.424
#>  3     0.0759      0.158
#>  4     0.753       0.792
#>  5     0.864       0.284
#>  6     0.548       0.921
#>  7     0.832       0.519
#>  8     0.719       0.432
#>  9     0.893       0.916
#> 10     0.591       0.519
#> 
#> 
#> $detected$sample2
#> $detected$sample2$gene2
#> # A tibble: 20 × 2
#>    x_location y_location
#>         <dbl>      <dbl>
#>  1     0.0619     0.0388
#>  2     0.411      0.403 
#>  3     0.568      0.244 
#>  4     0.548      0.407 
#>  5     0.645      0.598 
#>  6     0.834      0.191 
#>  7     0.551      0.263 
#>  8     0.634      0.571 
#>  9     0.278      0.268 
#> 10     0.540      0.233 
#> 11     0.111      0.230 
#> 12     0.182      0.129 
#> 13     0.164      0.318 
#> 14     0.953      0.634 
#> 15     0.698      0.979 
#> 16     0.448      0.0365
#> 17     0.0141     0.189 
#> 18     0.587      0.0332
#> 19     0.616      0.289 
#> 20     0.139      0.287 
#> 
#> 
#>