Read in csv boundary information and convert to ME list format.

This function reads in csv boundary files and converts them to the ME list format, so that they can be added to an ME object later on. To account for different coordinate scales possible being used by the boundary versus transcript information, this function scales the coordinate values of the boundaries to match the unit of the detected transcript locations. The various arguments offer flexibility to standardise data from different molecule-based ST technologies into the ME list format.

Usage

readBoundaries(
  data_dir,
  pattern = NULL,
  n_samples = NULL,
  segment_id_col = NULL,
  x_col = NULL,
  y_col = NULL,
  keep_cols = "essential",
  boundaries_assay = NULL,
  scale_factor_vector = 1
)

Arguments

data_dir

Path of the directory containing the boundary csv files.

pattern

Character string specifying the unique pattern with which to identify the files of interest in the directory. This is useful to work with multiple samples. Defaults to NULL.

n_samples

Integer indicating the number of samples. Defaults to NULL.

segment_id_col

Character string specifying the name of the column containing the sample id. Defaults to NULL.

x_col

Character string specifying the name of the column containing the x coordinates of the vertices defining the boundaries. Defaults to NULL.

y_col

Character string specifying the name of the column containing the y coordinates of the vertices defining the boundaries. Defaults to NULL.

keep_cols

Character string specifying which columns to keep. Defaults to "essential". The other option is to select "all", or custom columns by specifying their names in a vector.

boundaries_assay

Character string specifies. Defaults to NULL.

scale_factor_vector

Vector containing the scale factor/s with which to change the coordinate data from pixel to micron. It can be either a single integer, or multiple scale factors for the different samples. The default value is 1.

Examples

  repo_dir <- system.file("extdata", package = "MoleculeExperiment")
  nuclei_ls <- readBoundaries(data_dir = repo_dir,
                              pattern = "nucleus_boundaries.csv",
                              n_samples = 2,
                              segment_id_col = "cell_id",
                              x_col = "vertex_x",
                              y_col = "vertex_y",
                              keep_cols = "essential",
                              boundaries_assay = "nucleus",
                              scale_factor_vector = 0.2125)
  nuclei_ls
#> $nucleus
#> $nucleus$Xenium_V1_FF_Mouse_Brain_MultiSection_1_outs
#> $nucleus$Xenium_V1_FF_Mouse_Brain_MultiSection_1_outs$`67500`
#> # A tibble: 13 × 2
#>    x_location y_location
#>         <dbl>      <dbl>
#>  1      1041.      1362.
#>  2      1041.      1362.
#>  3      1041.      1362.
#>  4      1041.      1363.
#>  5      1041.      1363.
#>  6      1041.      1363.
#>  7      1042.      1363.
#>  8      1042.      1363.
#>  9      1042.      1363.
#> 10      1042.      1362.
#> 11      1042.      1362.
#> 12      1042.      1362.
#> 13      1041.      1362.
#> 
#> $nucleus$Xenium_V1_FF_Mouse_Brain_MultiSection_1_outs$`67512`
#> # A tibble: 13 × 2
#>    x_location y_location
#>         <dbl>      <dbl>
#>  1      1043.      1361.
#>  2      1043.      1362.
#>  3      1043.      1362.
#>  4      1043.      1363.
#>  5      1043.      1363.
#>  6      1044.      1363.
#>  7      1044.      1363.
#>  8      1044.      1363.
#>  9      1044.      1362.
#> 10      1044.      1362.
#> 11      1044.      1362.
#> 12      1043.      1361.
#> 13      1043.      1361.
#> 
#> $nucleus$Xenium_V1_FF_Mouse_Brain_MultiSection_1_outs$`67515`
#> # A tibble: 13 × 2
#>    x_location y_location
#>         <dbl>      <dbl>
#>  1      1044.      1361.
#>  2      1043.      1361.
#>  3      1043.      1361.
#>  4      1044.      1362.
#>  5      1044.      1362.
#>  6      1044.      1363.
#>  7      1045.      1362.
#>  8      1044.      1362.
#>  9      1045.      1362.
#> 10      1045.      1362.
#> 11      1044.      1361.
#> 12      1044.      1361.
#> 13      1044.      1361.
#> 
#> $nucleus$Xenium_V1_FF_Mouse_Brain_MultiSection_1_outs$`67521`
#> # A tibble: 13 × 2
#>    x_location y_location
#>         <dbl>      <dbl>
#>  1      1045.      1362.
#>  2      1045.      1362.
#>  3      1045.      1362.
#>  4      1044.      1363.
#>  5      1044.      1364.
#>  6      1045.      1364.
#>  7      1045.      1364.
#>  8      1045.      1364.
#>  9      1045.      1363.
#> 10      1045.      1363.
#> 11      1045.      1362.
#> 12      1045.      1362.
#> 13      1045.      1362.
#> 
#> $nucleus$Xenium_V1_FF_Mouse_Brain_MultiSection_1_outs$`67527`
#> # A tibble: 13 × 2
#>    x_location y_location
#>         <dbl>      <dbl>
#>  1      1046.      1362.
#>  2      1046.      1363.
#>  3      1045.      1363.
#>  4      1045.      1364.
#>  5      1045.      1364.
#>  6      1046.      1364.
#>  7      1046.      1364.
#>  8      1046.      1364.
#>  9      1046.      1364.
#> 10      1046.      1363.
#> 11      1046.      1363.
#> 12      1046.      1362.
#> 13      1046.      1362.
#> 
#> 
#> $nucleus$Xenium_V1_FF_Mouse_Brain_MultiSection_2_outs
#> $nucleus$Xenium_V1_FF_Mouse_Brain_MultiSection_2_outs$`65044`
#> # A tibble: 13 × 2
#>    x_location y_location
#>         <dbl>      <dbl>
#>  1      1042.      1364.
#>  2      1042.      1364.
#>  3      1042.      1365.
#>  4      1042.      1365.
#>  5      1042.      1365.
#>  6      1042.      1365.
#>  7      1043.      1365.
#>  8      1043.      1365.
#>  9      1043.      1365.
#> 10      1043.      1365.
#> 11      1043.      1364.
#> 12      1042.      1364.
#> 13      1042.      1364.
#> 
#> $nucleus$Xenium_V1_FF_Mouse_Brain_MultiSection_2_outs$`65051`
#> # A tibble: 13 × 2
#>    x_location y_location
#>         <dbl>      <dbl>
#>  1      1045.      1363.
#>  2      1045.      1364.
#>  3      1045.      1365.
#>  4      1045.      1365.
#>  5      1045.      1366.
#>  6      1045.      1366.
#>  7      1045.      1366.
#>  8      1046.      1366.
#>  9      1046.      1366.
#> 10      1046.      1365.
#> 11      1046.      1364.
#> 12      1046.      1363.
#> 13      1045.      1363.
#> 
#> $nucleus$Xenium_V1_FF_Mouse_Brain_MultiSection_2_outs$`65055`
#> # A tibble: 13 × 2
#>    x_location y_location
#>         <dbl>      <dbl>
#>  1      1043.      1362.
#>  2      1043.      1362.
#>  3      1042.      1362.
#>  4      1042.      1363.
#>  5      1042.      1363.
#>  6      1042.      1364.
#>  7      1043.      1364.
#>  8      1043.      1364.
#>  9      1044.      1363.
#> 10      1044.      1363.
#> 11      1044.      1362.
#> 12      1043.      1362.
#> 13      1043.      1362.
#> 
#> $nucleus$Xenium_V1_FF_Mouse_Brain_MultiSection_2_outs$`65063`
#> # A tibble: 13 × 2
#>    x_location y_location
#>         <dbl>      <dbl>
#>  1      1046.      1363.
#>  2      1046.      1363.
#>  3      1046.      1363.
#>  4      1045.      1363.
#>  5      1046.      1363.
#>  6      1046.      1364.
#>  7      1046.      1363.
#>  8      1046.      1363.
#>  9      1047.      1363.
#> 10      1047.      1363.
#> 11      1047.      1363.
#> 12      1047.      1363.
#> 13      1046.      1363.
#> 
#> $nucleus$Xenium_V1_FF_Mouse_Brain_MultiSection_2_outs$`65064`
#> # A tibble: 13 × 2
#>    x_location y_location
#>         <dbl>      <dbl>
#>  1      1041.      1361.
#>  2      1040.      1361.
#>  3      1040.      1362.
#>  4      1040.      1362.
#>  5      1040.      1362.
#>  6      1040.      1363.
#>  7      1041.      1363.
#>  8      1041.      1362.
#>  9      1041.      1362.
#> 10      1042.      1361.
#> 11      1041.      1361.
#> 12      1041.      1361.
#> 13      1041.      1361.
#> 
#> 
#>