Read in Xenium data into a MoleculeExperiment object

Function to read in, and standardise, Xenium output data into an ME object. Detected transcripts files are required. Additionally, it is also possible to read in boundary files ("cell", "nuclei", or both). This function is a wrapper around readMolecules and readBoundaries functions.

Usage

readXenium(dataDir, keepCols = "essential", addBoundaries = "cell")

Arguments

dataDir

Character string specifying the directory with the xenium output files.

keepCols

Vector of characters specifying the columns of interest from the transcripts file and boundaries file. Can be "all" or "essential". "essential" selects columns with gene names, x and y locations in the transcripts file; "essential" selects columns with cell ids, and x and y locations for the vertices defining the boundaries in the boundaries file.

addBoundaries

Vector with which to specify the names of the boundary assays to be added to the me object. Can be "cell", "nucleus", both, or NULL. The latter will lead to the creation of a simple ME object with just the molecules slot filled.

Value

A MoleculeExperiment object containing xenium data.

Examples

repoDir <- system.file("extdata", package = "MoleculeExperiment")
repoDir <- paste0(repoDir, "/xenium_V1_FF_Mouse_Brain")

me <- readXenium(repoDir,
                  keepCols = "essential")
me
#> class:  MoleculeExperiment 
#> 2 samples: sample1 sample2 
#> 
#> @molecules contents: 
#> -detected assay:
#> 178 unique features across all samples in assay detected: 2010300C02Rik Acsbg1
#> Adamts2 Adamtsl1 Angpt1 Aqp4...
#> 869.5 molecules on average across all samples in assay detected
#> Location range across all samples in assay "detected": [4900,4919.98] x [6400.02,6420]
#> 
#> @boundaries contents:
#> -cell:
#> 7 unique segment IDs: 67500 67512 67515 67521 67527 ...